+++ src/vs/platform/quickinput/browser/media/quickInput.css
