+++ src/vs/platform/quickinput/browser/quickInputBox.ts
