+++ src/vs/platform/quickinput/browser/quickInputUtils.ts
